code/requirements.txt
torch==2.0.0+cu118 torchvision==0.15.1+cu118 torchaudio==2.0.1 --index-url https://download.pytorch.org/whl/cu118
nested_dict
matplotlib
seaborn
numpy==1.21.2
scipy==1.7.3
pandas==1.1.5
tensorboard==2.4.1
flake8==3.3.0
pep8-naming==0.4.1
scikit-learn==0.24.0
pytest==3.2.1
ray
tabulate==0.8.9
hyperopt==0.2.5
protobuf==3.20.0
tqdm==4.60.0
fastdtw==0.3.4
networkx==2.5.1
gensim==4.0.1
nltk==3.2.4
statsmodels==0.13.1
tslearn==0.5.2
pytorch-lightning

